resume_website_restapi/pyproject.toml
[tool.poetry]
name = "resume-website-restapi"
version = "0.1.0"
description = ""
authors = ["Vlad_the_programmer <vladyslavklymchuk@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
argcomplete = "^1.8.1"
asgiref = "^3.9.1"
attrs = "^21.2.0"
Automat = "^20.2.0"
Babel = "^2.8.0"
bcrypt = "^3.2.0"
black = "^25.1.0"
blinker = "^1.4"
certifi = "^2025.8.3"
cffi = "^1.17.1"
chardet = "^4.0.0"
charset-normalizer = "^3.4.3"
click = "^8.0.3"
colorama = "^0.4.4"
configobj = "^5.0.6"
constantly = "^15.1.0"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
crispy-bootstrap5 = "^2025.6"
cryptography = "^45.0.6"
cssselect = "^1.2.0"
defusedxml = "^0.7.1"
distlib = "^0.3.4"
distro = "^1.7.0"
dj-rest-auth = {extras = ["with-social"], version = "^7.0.1"}
Django = "^5.2.5"
django-allauth = "^65.10.0"
django-countries = "^7.6.1"
django-crispy-forms = "^2.4"
django-filter = "^25.1"
django-jazzmin = "^3.0.1"
django-templated-mail = "^1.1.1"
djangorestframework = "^3.16.1"
djangorestframework_simplejwt = "^5.5.1"
filelock = "^3.6.0"
flake8 = "^7.2.0"
greenlet = "^3.1.1"
httplib2 = "^0.20.2"
hyperlink = "^21.0.0"
idna = "^3.10"
importlib-metadata = "^4.6.4"
incremental = "^21.3.0"
itypes = "^1.2.0"
Jinja2 = "^3.1.6"
jsonpatch = "^1.32"
jsonpointer = "^2.0"
jsonschema = "^3.2.0"
keyring = "^23.5.0"
livereload = "^2.6.3"
lxml = "^4.9.1"
Markdown = "^3.4.1"
MarkupSafe = "^3.0.2"
mccabe = "^0.7.0"
mkdocs = "^1.1.2"
more-itertools = "^8.10.0"
mypy = "^1.16.1"
mypy_extensions = "^1.1.0"
oauthlib = "^3.3.1"
packaging = "^25.0"
pathspec = "^0.12.1"
pexpect = "^4.8.0"
Pillow = "^11.3.0"
pipenv = "^11.9.0"
pipx = "^1.0.0"
platformdirs = "^2.5.1"
psutil = "^5.9.0"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
ptyprocess = "^0.7.0"
pyasn1 = "^0.4.8"
pyasn1-modules = "^0.2.1"
pycodestyle = "^2.13.0"
pycparser = "^2.22"
pyflakes = "^3.3.2"
Pygments = "^2.11.2"
PyHamcrest = "^2.0.2"
pyinotify = "^0.9.6"
PyJWT = "^2.10.1"
PyNaCl = "^1.5.0"
pyOpenSSL = "^21.0.0"
pyparsing = "^2.4.7"
pyquery = "^1.4.3"
pyrsistent = "^0.18.1"
pyserial = "^3.5"
python-dotenv = "^1.0.1"
python-magic = "^0.4.24"
python3-openid = "^3.2.0"
pytz = "^2025.2"
PyYAML = "^6.0"
requests = "^2.32.4"
requests-oauthlib = "^2.0.0"
SecretStorage = "^3.3.1"
service-identity = "^18.1.0"
six = "^1.17.0"
social-auth-app-django = "^5.5.1"
social-auth-core = "^4.7.0"
SQLAlchemy = "^2.0.38"
sqlparse = "^0.5.3"
tomli = "^2.2.1"
tornado = "^6.1"
Twisted = "^22.1.0"
typing_extensions = "^4.14.1"
uritemplate = "^4.2.0"
urllib3 = "^2.5.0"
virtualenv = "^20.13.0"
virtualenv-clone = "^0.3.0"
zipp = "^1.0.0"
drf-spectacular = "^0.27.1"  # For Swagger/OpenAPI

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"